added style to country page
