--- src/components/Video.js
@@ -1,5 +1,0 @@
-export default function Video(props) {
-  return(
-    <video src={props.src} controls="true"></video>
-  )
-}
