+++ src/components/Widget/Widget.css
@@ -1,11 +1,12 @@
 .widget{
   position: absolute;
   top: 50px;
-  right: 300px;
+  right: 5.3%;
   background-color: rgba(0, 0, 0, 0.3);
   padding: 20px;
   border-radius: 8px;
   display: flex;
   flex-flow: column nowrap;
   row-gap: 20px;
+  z-index: 2;
 }
