+++ src/components/countru-page.component/Video.js
@@ -1,0 +1,7 @@
+export default function Video(props) {
+  return(
+    <div className="vidio__container">
+      <video src={props.src} controls="true"></video>
+    </div>
+  )
+}
