+++ src/components/countru-page.component/country-page.css
@@ -1,0 +1,84 @@
+.country-page-content__container {
+  width: 90%;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.vidio__container,
+.country-description__container,
+.carousel_wrapper,
+.leaflet-container,
+video {
+  width: 100%;
+}
+
+.country-name__header {
+  position: absolute;
+  top: 90vh;
+  width: 90%;
+
+  display: block;
+  padding: 20px;
+
+  background-color: black;
+  opacity: 0.5;
+
+  z-index: 16;
+}
+
+.country-description__container {
+  min-height: 90vh;
+  margin: 50px 0;
+  padding: 5px;
+
+  background-size: cover;
+}
+
+.country-description__content{
+  min-width: 300px;
+  width: 70%;
+  height: 88vh;
+  padding: 10px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+
+  background-color: rgba(000, 000, 000, 0.4);
+  border: 1px solid white;
+}
+.full-screen-map__container{
+  position: fixed;
+  top: 0;
+  left: 0;
+  height: 100vh;
+  width: 100vw;
+}
+
+.map__container{
+  position: relative;
+  height: 250px;
+  min-width: 300px;
+  width: 100%;
+}
+
+.leaflet-container {
+  width: 100%;
+  height: Calc(100% - 25px);
+  z-index: 5;
+
+  border: 1px solid black;
+  border-radius: 5px;
+}
+
+.map-toggler {
+  position: relative;
+}
+
+video {
+  height: 100%;
+  outline: none;
+}
