+++ src/components/main-page.component/CountryCard.js
